Exercise 7 and Phase 2, done
