+++ README.md
@@ -1,8 +1,9 @@
 # ip21_akademija_projekt1
 
 Application for accessing cryptocurrency price data, built step by step.
+
 At its current state, the App makes an API call, decodes json file and displays the value of selected crypto in the selected currency.
-It can also handle certain common errors.
+It can also handle certain common errors and now has basic MVC structure.
